Move parse_str_addr and ProtocolId to config
